+++ src/main/kotlin/io/github/rybalkinsd/kohttp/dsl/async/AsyncHttpGetDsl.kt
@@ -1,7 +1,7 @@
 package io.github.rybalkinsd.kohttp.dsl.async
 
 import io.github.rybalkinsd.kohttp.client.defaultHttpClient
-import io.github.rybalkinsd.kohttp.dsl.HttpGetContext
+import io.github.rybalkinsd.kohttp.dsl.context.HttpGetContext
 import io.github.rybalkinsd.kohttp.ext.suspendCall
 import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.Dispatchers.Unconfined
