+++ src/main/kotlin/io/github/rybalkinsd/kohttp/dsl/context/ParamContext.kt
@@ -1,0 +1,12 @@
+package io.github.rybalkinsd.kohttp.dsl.context
+
+@HttpDslMarker
+class ParamContext {
+    private val map: MutableMap<String, Any> = mutableMapOf()
+
+    infix fun String.to(v: Any) {
+        map[this] = v
+    }
+
+    internal fun forEach(action: (k: String, v: Any) -> Unit) = map.forEach(action)
+}
